pyproject.toml
[project]
name = "tpa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "google-genai>=1.56.0",
    "ipython>=9.8.0",
    "openai>=2.15.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.45",
    "babel>=2.17.0",
    "pycountry>=24.6.1",
    "pytest>=9.0.2",
]
